+++ CHANGELOG.md
@@ -7,6 +7,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 
+## [0.2.0] - 2018-03-18
+
+### Breaking changes
+- Breaking changes to config file, check the [latest config here](./posti.config.example.js).
+- In `postalcodes` table, renamed column `effectiveAt` to `entryIntoForceAt`.
+
+### Changed
+- Improvements in tests.
+
+
 ## [0.1.6] - 2018-03-18
 
 ### Added
@@ -49,19 +59,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.1.0] - 2018-03-17
 
+### Breaking changes
+- Breaking changes to config file, check the [latest config here](./posti.config.example.js).
+- Table `zipcode_changes` renamed to `postalcode_changes`.
+- Table `zipcodes` renamed to `postalcodes`.
+
 ### Added
 - Disclaimer in `README.md`.
 - Links to Finnish, English and Swedish Postal Code Services site, Terms of use and Frequently asked questions.
 - Babel polyfill in `src/posti.js`.
 
 ### Changed
-- Breaking changes to config file, check the [latest config here](./posti.config.example.js).
 - Major refactoring.
 - Updated dependencies.
 - Run tests in parallel instead in serially.
 - Coverage percentage in badge to contain `% Branch` instead of `% Funcs`.
-- Table `zipcode_changes` renamed to `postalcode_changes`.
-- Table `zipcodes` renamed to `postalcodes`.
 - `Stage-0` into `Stage-2`.
 
 ### Fixed